Added checks for user status

+++ app/controllers/bookings_controller.rb
@@ -4,43 +4,33 @@ class BookingsController < ApplicationController
   before_action :find_user
 
 
-  def missing_information
-    current_user.user_profile.first_name.empty? ||
-    current_user.user_profile.last_name.empty? ||
-    current_user.user_profile.dob.nil? ||
-    current_user.user_profile.gender.nil? ||
-    current_user.user_profile.picture_file_name.nil? ||
-    current_user.user_profile.native_lang.empty? ||
-    current_user.user_profile.first_lang.empty? ||
-    current_user.user_profile.second_lang.empty?
-    
-  
-  end
-
   def index
-    if missing_information
-          flash[:alert] = "You can only view your booking if you have filled out your user profile"
-      redirect_to  edit_user_profile_path(current_user)
-    end
-
-    @bookings = Booking.all
-    @userbookings = []
+    return redirect_to_sign_in if missing_information
     @user = current_user
     if @user
+      @bookings = Booking.all
+      @userbookings = []
+      @user = current_user
       @bookings.each do |x|
         if x.user_id == @user.id || x.student_id == @user.id
           @userbookings << x
         end 
       end
     end
-   @userbookings
+    @userbookings
   end
 
   def new
+    return redirect_to_sign_in if missing_information
     @booking = Booking.new(user_id: @user.id)
   end
 
   def create
+    return redirect_to_sign_in if missing_information
+    if @user.id = current_user
+      flash[:alert] = "You cannot invite yourself!"
+      return redirect_to '/'
+    end
     @booking =  Booking.new(params[:booking].permit(:user_id, :start_time, :length, :topic_id))
     @booking.lang1 = @user.native_lang
     @booking.user_id = @user.id
@@ -62,10 +52,12 @@ class BookingsController < ApplicationController
   end
 
   def show
+    return redirect_to_sign_in if missing_information
     @booking = Booking.find(params[:id])
   end
 
   def destroy
+    return redirect_to_sign_in if missing_information
     @booking = Booking.find(params[:id]).destroy
     if @booking.destroy
       flash[:notice] = "Booking: #{@booking.start_time.strftime('%e %b %Y %H:%M%p')} to #{@booking.end_time.strftime('%e %b %Y %H:%M%p')} deleted"
@@ -76,11 +68,13 @@ class BookingsController < ApplicationController
   end
 
   def edit
+    return redirect_to_sign_in if missing_information
     @booking = Booking.find(params[:id])
     @user = current_user
   end
 
   def update
+    return redirect_to_sign_in if missing_information
     @booking = Booking.find(params[:id])
     # @booking.user = @user
